// src/ai-contract.generated.ts
export const AI_CONTRACT_GENERATED: AiContractGeneratedData = {
  "headerDeclarations": [
    {
      "name": "SeedInput",
      "declarationLines": [
        "export type SeedInput = string | number | bigint;"
      ]
    },
// ...
} as const;